Feature/advanced discovery

+++ prism-search-service/src/main/java/com/smile/prism/search/api/SearchResponse.java
@@ -4,11 +4,14 @@ import com.smile.prism.search.model.EventDocument;
 import java.util.List;
 
 /**
- * Immutable Search Response.
- * Leveraging Java 21 Records for efficient memory management in high-throughput scenarios.
+ * Standardized Search Response for the Prism Discovery Engine.
+ *
+ * @param totalHits The total number of documents matching the query in Elasticsearch.
+ * @param results The specific page of results.
+ * @param status Status of the search operation (e.g., SUCCESS, PARTIAL_RESULTS).
  */
 public record SearchResponse(
-        int totalHits,
+        long totalHits,
         List<EventDocument> results,
         String status
 ) {}
